Module Server for Assay Specification

This defines the server part for the assay specification.

assaySpecServer(inputId, assays, exclude_assays = character())

Arguments

inputId
(string)
the ID used to call the module input.

assays
(reactive character)
available assays in the currently selected experiment.

exclude_assays
(character)
names of the assays which should not be included in choices in the teal module.

Value

The chosen assay as a reactive string.

See also

Examples

ui <- function(id,
               datasets) {
  ns <- NS(id)
  teal.devel::standard_layout(
    encoding = div(
      experimentSpecInput(
        ns("experiment"),
        datasets,
        "MAE"
      ),
      assaySpecInput(
        ns("assay"),
        label_assays = "Please choose assay"
      )
    ),
    output = textOutput(ns("result"))
  )
}

server <- function(input,
                   output,
                   session,
                   datasets) {
  experiment <- experimentSpecServer(
    "experiment",
    datasets,
    "MAE"
  )
  assay <- assaySpecServer(
    "assay",
    experiment$assays,
    exclude_assays = c("counts", "cpm", "tpm", "bla")
  )
  output$result <- renderPrint({
    assay()
  })
}

my_app <- function() {
  mae <- hermes::multi_assay_experiment
  mae_name <- "MAE"
  mae_data <- dataset(mae_name, mae)
  data <- teal_data(mae_data)
  app <- init(
    data = data,
    modules = root_modules(
      module(
        label = "assaySpec example",
        server = server,
        ui = ui,
        filters = "all"
      )
    )
  )
  shinyApp(app$ui, app$server)
}
if (interactive()) {
  my_app()
}