Data Preprocessing for KM Module

A function to help with merging of MAE to ADTTE for use with g_km.

h_km_mae_to_adtte(
  adtte,
  mae,
  gene_var,
  experiment_name = "hd1",
  assay_name = "counts"
)

Arguments

adtte
(data frame)
an adtte dataset.

mae
(MultiAssayExperiment)
contains AnyHermesData objects.

gene_var
(character)
a gene(s) of interest.

experiment_name
(string)
the desired HermesData to use.

assay_name
(string)
the assay to define the groups.

Value

A data frame containing all columns/rows from adtte that match by USUBJID with the row names of the MAE and have the gene samples available in the given experiment. The attributes sample_id and gene_cols contain the column names for the sample ID and gene columns.

Note

We require that each patient only has one sample.

The final gene column names can start with a different string than the original gene IDs, in particular white space, dots and colons are removed, see tern::make_names() for details.

Examples

mae <- hermes::multi_assay_experiment
adtte <- scda::synthetic_cdisc_data("rcd_2021_07_07")$adtte %>%
  dplyr::mutate(CNSR = as.logical(CNSR))

new_adtte <- h_km_mae_to_adtte(
  adtte,
  mae,
  gene_var = "GeneID:1820",
  experiment_name = "hd2"
)

#> Registered S3 method overwritten by 'tern':
#>   method   from 
#>   tidy.glm broom

new_adtte2 <- h_km_mae_to_adtte(
  adtte,
  mae,
  gene_var = c("GeneID:1820", "GeneID:94115"),
  experiment_name = "hd2"
)