Teal Module for RNA-seq Barplot

This module provides an interactive barplot for RNA-seq gene expression analysis.

Usage

tm_g_barplot(
  label,
  mae_name,
  exclude_assays = character(),
  summary_funs = list(Mean = colMeans, Median = matrixStats::colMedians, Max =
    matrixStats::colMaxs),
  pre_output = NULL,
  post_output = NULL
)

ui_g_barplot(id, datasets, mae_name, summary_funs, pre_output, post_output)

srv_g_barplot(id, datasets, reporter, mae_name, exclude_assays, summary_funs)

sample_tm_g_barplot()

Arguments

label

(string)
menu item label of the module in the teal app.

mae_name

(string)
name of the MAE data used in the teal module.

exclude_assays

(character)
names of the assays which should not be included in choices in the teal module.

summary_funs

(named list of functions or NULL)
functions which can be used in the the gene signatures. For modules that support also multiple genes without summary, NULL can be included to not summarize the genes but provide all of them.

pre_output

(shiny.tag or NULL)
placed before the output to put the output into context (for example a title).

post_output

(shiny.tag or NULL)
placed after the output to put the output into context (for example the shiny::helpText() elements can be useful).

id

(string) the shiny module id.

datasets

(Datasets)
teal specific argument which is automatically passed to the UI and server functions, holding all the data sets provided in the app initialization.

reporter

(Reporter) object

Value

Shiny module to be used in the teal app.

Functions

ui_g_barplot(): sets up the user interface.

srv_g_barplot(): sets up the server with reactive graph.

sample_tm_g_barplot(): sample module function.

Examples

mae <- hermes::multi_assay_experiment
mae_data <- dataset("MAE", mae)
data <- teal_data(mae_data)
app <- init(
  data = data,
  modules = modules(
    tm_g_barplot(
      label = "barplot",
      mae_name = "MAE"
    )
  )
)
#> [INFO] 2022-10-17 14:23:49.5265 pid:2146 token:[] teal.modules.hermes Initializing tm_g_barplot
if (interactive()) {
  shinyApp(app$ui, app$server)
}

# Alternatively you can run the sample module with this function call:
if (interactive()) {
  sample_tm_g_barplot()
}