Teal Module for Kaplan-Meier Plot

This teal module produces a grid style Kaplan-Meier plot for data with ADaM structure.

tm_g_km(
  label,
  adtte_name,
  mae_name,
  strata_var = NULL,
  pre_output = NULL,
  post_output = NULL
)

ui_g_km(
  id,
  datasets,
  adtte_name,
  mae_name,
  strata_var,
  pre_output,
  post_output
)

srv_g_km(input, output, session, datasets, adtte_name, mae_name, strata_var)

sample_tm_g_km()

Arguments

label
(string)
menu item label of the module in the teal app.

adtte_name
(string)
name of the ADTTE dataset.

mae_name
(string)
name of the MAE data used in the teal module.

strata_var
(character)
names of the stratification variables.

pre_output
(shiny.tag or NULL)
placed before the output to put the output into context (for example a title).

post_output
(shiny.tag or NULL)
placed after the output to put the output into context (for example the shiny::helpText() elements can be useful).

id
(string)
the ID used to call the teal module's UI function.

datasets
(Datasets)
teal specific argument which is automatically passed to the UI and server functions, holding all the data sets provided in the app initialization.

input
(ReactiveValues)
the session's input object.

output
(shinyoutput)
the session's output object.

session
(ShinySession)
the session object.

Value

Shiny module to be used in the teal app.

Functions

ui_g_km: sets up the user interface.

srv_g_km: sets up the user interface.

sample_tm_g_km: sample module function.

Examples

mae <- hermes::multi_assay_experiment
adtte <- scda::synthetic_cdisc_data("rcd_2021_07_07")$adtte %>%
  dplyr::mutate(CNSR = as.logical(CNSR))

data <- teal_data(
  dataset(
    "ADTTE",
    adtte,
    code = 'adtte <- radtte(cached = TRUE) %>% mutate(CNSR = as.logical(CNSR))'
  ),
  dataset("mae", mae)
)

modules <- root_modules(
  tm_g_km(
    label = "KM PLOT",
    adtte_name = "ADTTE",
    mae_name = "mae",
    strata_var = c("SEX", "STRATA1")
  )
)

app <- init(
  data = data,
  modules = modules
)

if (FALSE) {
shinyApp(ui = app$ui, server = app$server)
}

if (FALSE) {
# Alternatively you can run the sample module with this function call:
sample_tm_g_km()
}