Teal Module for RNA-seq Quality Control

This module adds quality flags, filters by genes and/or samples, normalizes AnyHermesData objects and provides interactive plots for RNA-seq gene expression quality control.

tm_g_quality(
  label,
  mae_name,
  exclude_assays = character(),
  pre_output = NULL,
  post_output = NULL
)

ui_g_quality(id, datasets, mae_name, pre_output, post_output)

srv_g_quality(input, output, session, datasets, mae_name, exclude_assays)

sample_tm_g_quality()

Arguments

label
(string)
menu item label of the module in the teal app.

mae_name
(string)
name of the MAE data used in the teal module.

exclude_assays
(character)
names of the assays which should not be included in choices in the teal module.

pre_output
(shiny.tag or NULL)
placed before the output to put the output into context (for example a title).

post_output
(shiny.tag or NULL)
placed after the output to put the output into context (for example the shiny::helpText() elements can be useful).

id
(string)
the ID used to call the teal module's UI function.

datasets
(Datasets)
teal specific argument which is automatically passed to the UI and server functions, holding all the data sets provided in the app initialization.

input
(ReactiveValues)
the session's input object.

output
(shinyoutput)
the session's output object.

session
(ShinySession)
the session object.

Value

Shiny module to be used in the teal app.

Functions

ui_g_quality: sets up the user interface.

srv_g_quality: sets up the server with reactive graphs.

sample_tm_g_quality: sample module function.

Examples

mae <- hermes::multi_assay_experiment
mae_data <- dataset("MAE", mae)
data <- teal_data(mae_data)
app <- init(
  data = data,
  modules = root_modules(
    tm_g_quality(
      label = "Quality Control",
      mae_name = "MAE"
    )
  )
)
if (interactive()) {
  shinyApp(app$ui, app$server)
}

# Alternatively you can run the sample module with this function call:
if (interactive()) {
  sample_tm_g_quality()
}